// src/app/types/shipping/methods/omniva/omniva.ts
export interface OmnivaInterface {
    zip: string
    name: string
    type: string
    a0_name: string
    a1_name: string
    a2_name: string
    a3_name: string
    a4_name: string
    a5_name: string
    a6_name: string
    a7_name: string
    a8_name: string
    x_coordinate: string
    y_coordinate: string
    service_hours: string
    temp_service_hours: string
    temp_service_hours_until: string
    temp_service_hours_2: string
    temp_service_hours_2_until: string
    comment_est: string
    comment_eng: string
    comment_rus: string
    comment_lav: string
    comment_lit: string
    modified: string
}

export interface DeliveryCountryInterface {
    country_code: string;
    country_name: string;
}